chore!: rename testutil to util
Motivation: "By convention, the last element of the package path is the package
name"

Source: https://go.dev/blog/package-names
