+++ test/util/direct_tx_gen.go
@@ -1,4 +1,4 @@
-package testutil
+package util
 
 import (
 	"math/rand"
