+++ test/util/test_app.go
@@ -1,4 +1,4 @@
-package testutil
+package util
 
 import (
 	"encoding/json"
